+++ patches/src/main/kotlin/app/morphe/patches/youtube/layout/buttons/overlay/HidePlayerOverlayButtonsPatch.kt
@@ -6,6 +6,8 @@ import app.morphe.patcher.extensions.InstructionExtensions.addInstructionsWithLa
 import app.morphe.patcher.extensions.InstructionExtensions.getInstruction
 import app.morphe.patcher.patch.bytecodePatch
 import app.morphe.patcher.util.smali.ExternalLabel
+import app.morphe.patches.all.misc.resources.ResourceType
+import app.morphe.patches.all.misc.resources.getResourceId
 import app.morphe.patches.all.misc.resources.resourceMappingPatch
 import app.morphe.patches.shared.misc.settings.preference.SwitchPreference
 import app.morphe.patches.youtube.misc.extension.sharedExtensionPatch
@@ -52,12 +54,18 @@ val hidePlayerOverlayButtonsPatch = bytecodePatch(
             SwitchPreference("morphe_hide_fullscreen_button"),
             SwitchPreference("morphe_hide_player_control_buttons_background"),
             SwitchPreference("morphe_hide_player_previous_next_buttons"),
+            SwitchPreference("morphe_hide_settings_button"),
         )
 
-        // region Hide player next/previous button.
+        // region Hide player previous/next & settings button.
 
         LayoutConstructorFingerprint.let {
             it.clearMatch() // Fingerprint is shared with other patches.
+
+            // Verify resources exist.
+            getResourceId(ResourceType.ID, "player_control_next_button_touch_area")
+            getResourceId(ResourceType.ID, "player_control_previous_button_touch_area")
+            getResourceId(ResourceType.ID, "player_overflow_button")
 
             it.method.apply {
                 val insertIndex = it.instructionMatches.last().index
@@ -67,6 +75,12 @@ val hidePlayerOverlayButtonsPatch = bytecodePatch(
                     insertIndex,
                     "invoke-static { v$viewRegister }, $EXTENSION_CLASS" +
                             "->hidePreviousNextButtons(Landroid/view/View;)V",
+                )
+
+                addInstruction(
+                    insertIndex,
+                    "invoke-static { v$viewRegister }, $EXTENSION_CLASS" +
+                            "->hideSettingsButton(Landroid/view/View;)V",
                 )
             }
         }