+++ patches/src/main/kotlin/app/morphe/patches/youtube/layout/hide/general/HideLayoutComponentsPatch.kt
@@ -125,9 +125,9 @@ val hideLayoutComponentsPatch = bytecodePatch(
                     SwitchPreference("morphe_hide_info_cards_section"),
                     SwitchPreference("morphe_hide_key_concepts_section"),
                     SwitchPreference("morphe_hide_music_section"),
-                    SwitchPreference("morphe_hide_subscribe_button"),
-                    SwitchPreference("morphe_hide_transcript_section"),
                     SwitchPreference("morphe_hide_quizzes_section"),
+                    SwitchPreference("morphe_hide_subscribe_button"),
+                    SwitchPreference("morphe_hide_transcript_section")
                 ),
             ),
             PreferenceScreenPreference(